Added my-blogs page and a buggy delete method
The script tag will only work for the delete 'button' (on click event listener) on the first blog.  It doesn't look like it, but it will....  It will call the redirect and delete methods...  However, it only deletes the blog from the Blog cluster on the mongoDB.  It doesn't delete it from the array of the account it's on... So, it seems like clicking the first delete does nothing, but you can't find it on the /browse page, and clicking the link from the /my-blogs page will cause error.